+++ functions/apis/index.js
@@ -2,7 +2,12 @@
  * @file ./functions/apis/index.js
  * @description 公開するCallable APIを集約してexportするエントリーポイントです。
  *****************************************************************************/
+export { changeAdminUser } from "./changeAdminUser.js";
 export { checkEmailAvailability } from "./checkEmailAvailability.js";
 export { checkEmailAvailabilityGlobal } from "./checkEmailAvailabilityGlobal.js";
+export { checkUserPreRegistration } from "./checkUserPreRegistration.js";
+export { createAdminAccount } from "./createAdminAccount.js";
+export { disableUser, enableUser } from "./changeUserEnabledState.js";
 export { rebuildAllHistories } from "./rebuildAllHistories.js";
 export { rebuildSecurityReportIndexes } from "./rebuildSecurityReportIndexes.js";
+export { setupUserAccount } from "./setupUserAccount.js";